refactor(motion): rename seam_harness -> seam_test_harness
The module is test-support only (gated behind #![cfg(any(test,
feature = "test-support"))]); name it so its test-only role is obvious at
the call site and in the source tree.
