+++ rust/motion-engine/src/seam_test_harness.rs
