+++ rust/motion-engine/tests/seam_schedule_fuzz.rs
@@ -1,6 +1,6 @@
 use std::sync::OnceLock;
 
-use _motion_engine::seam_harness::{
+use _motion_engine::seam_test_harness::{
     Cadence, CommitSchedule, Move, default_stream_config, parse_gcode_to_moves, run_moves,
 };
 use proptest::prelude::*;
